refactor: put type translation functions in own file
