+++ react-components/src/components/NodeCacheProvider/types.ts
@@ -5,8 +5,6 @@ import { type Node3D } from '@cognite/sdk';
 import { type EdgeItem, type DmsUniqueIdentifier, type Source } from '../../utilities/FdmSDK';
 import { type InModel3dEdgeProperties } from '../../utilities/globalDataModels';
 
-import { split } from 'lodash';
-
 export type Fdm3dNodeData = { fdmId: DmsUniqueIdentifier; view: Source; cadNode: Node3D };
 export type FdmCadEdge = EdgeItem<InModel3dEdgeProperties>;
 export type FdmEdgeWithNode = { edge: FdmCadEdge; node: Node3D };
@@ -24,24 +22,3 @@ export type FdmKey = `${string}/${string}`;
 export type ModelNodeIdKey = `${ModelId}/${RevisionId}/${NodeId}`;
 
 export type ModelRevisionToEdgeMap = Map<ModelRevisionKey, FdmEdgeWithNode[]>;
-
-export function createModelRevisionKey(modelId: number, revisionId: number): ModelRevisionKey {
-  return `${modelId}/${revisionId}`;
-}
-
-export function revisionKeyToIds(revisionKey: ModelRevisionKey): [number, number] {
-  const components = split(revisionKey, '/');
-  return [Number(components[0]), Number(components[1])];
-}
-
-export function createModelNodeIdKey(
-  modelId: number,
-  revisionId: number,
-  nodeId: number
-): ModelNodeIdKey {
-  return `${modelId}/${revisionId}/${nodeId}`;
-}
-
-export function createFdmKey(spaceId: string, externalId: string): FdmKey {
-  return `${spaceId}/${externalId}`;
-}
